pyproject.toml
[project]
name = "ClouderaAIAgentStudio"
version = "0.0.1"
description = "Cloudera AI Agent Studio"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "arize-phoenix==8.20.0",
    "arize-phoenix-otel==0.9.0",
    "crewai==0.119.0",
    "gql==3.5.0",
    "grpcio==1.71.0",
    "grpcio-tools==1.71.0",
    "mcp==1.12.2",
    "mcpadapt==0.1.11",
    "mcpadapt[crewai]==0.1.11",
    "openinference-instrumentation-crewai==0.1.7",
    "openinference-instrumentation-langchain>=0.1.43",
    "openinference-instrumentation-litellm==0.1.14",
    "opentelemetry-api==1.31.1",
    "protobuf==5.29.4",
    "psutil==7.0.0",
    "pydantic==2.11.1",
    "pysqlite3-binary==0.5.4; platform_system == 'Linux'",
    "sqlalchemy==2.0.40",
    "uv==0.5.29",
    "alembic==1.15.2",
    "kombu==5.5.2",
    "openinference-instrumentation-langchain>=0.1.43",
    "boto3>=1.40.3",
    "botocore>=1.40.3",
]

[dependency-groups]
dev = [
    "autoflake==2.3.1",
    "debugpy==1.8.12",
    "pytest==8.2.2",
    "pytest-cov==5.0.0",
    "ruff==0.9.4",
]

[tool.ruff]
line-length = 120
